feat: make Linkify component reusable
